Added MethodOverloading.java Ver1.1

+++ src/ncku/unclethree/MethodOverloading.java
@@ -1,7 +1,5 @@
 package ncku.unclethree;
 
-import java.text.DecimalFormat;
-
 public class MethodOverloading {
     public static double calcFeetAndInchesToCentimeters(double feet, double inches) {
         if (feet < 0 || inches < 0 || inches > 12) {
@@ -10,12 +8,12 @@ public class MethodOverloading {
             return feet * 30.48 + inches * 2.54;
         }
     }
-    public static double calcFeetAndInchesToCentimeters(double inches){
+
+    public static double calcFeetAndInchesToCentimeters(double inches) {
         if (inches < 0) {
             return -1;
-        }
-        else{
-             return calcFeetAndInchesToCentimeters((int)inches/12 , (int) inches%12);
+        } else {
+            return calcFeetAndInchesToCentimeters((int) (inches / 12), inches % 12);
         }
     }
 }
